# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Authorazation",
        "title": "Authorazation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "paths": {
        "/user/login": {
            "post": {
                "description": "it generates new acces and refresh tokens",
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "user name and password",
                        "name": "userinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "accestoken and refreshtoken",
                        "schema": {
                            "$ref": "#/definitions/authentication.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid date",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error while reading from server",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/refresh": {
            "post": {
                "description": "Checks refresh token. If valid, it returns a new access token.",
                "tags": [
                    "auth"
                ],
                "summary": "Check refresh token",
                "parameters": [
                    {
                        "description": "Refresh Token",
                        "name": "refreshToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.CheckRefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "accessToken",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "create new users",
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.UserDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authentication.ID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "authentication.AccessToken": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "authentication.CheckRefreshTokenRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "authentication.ID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "authentication.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "authentication.LoginResponse": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/authentication.AccessToken"
                },
                "refresh": {
                    "$ref": "#/definitions/authentication.RefreshToken"
                }
            }
        },
        "authentication.RefreshToken": {
            "type": "object",
            "properties": {
                "refreshtoken": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "authentication.UserDetails": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}